e2e/tools/validator/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "validator"
dynamic = ["version"]
description = ''
readme = "README.md"
requires-python = ">=3.11"
license = "Apache-2.0"
keywords = ["kepler", "kepler-model-server", "validator"]
authors = [
  { name = "Sunil Thaha", email = "sthaha@redhat.com" },
  { name = "Kaiyi Liu", email = "kaiyiliu21@gmail.com" },
]

classifiers = [
  "Development Status :: 4 - Beta",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: Implementation :: CPython",
  "Programming Language :: Python :: Implementation :: PyPy",
]
dependencies = [
  "click",
  "paramiko",
  "prometheus-api-client",
	"pyyaml",
	"numpy",
]

[project.scripts]
validator = "validator.cli:validator"

[tool.hatch.version]
path = "src/validator/__about__.py"

[tool.hatch.envs.default]
python = "3.11" # latest supported python on RHEL-9
dependencies = [
  "coverage[toml]>=6.5",
  "pytest",
  "ipython",
  "ipdb",
]

[tool.hatch.envs.default.scripts]
test = "pytest {args:tests}"
test-cov = "coverage run -m pytest {args:tests}"
cov-report = [
  "- coverage combine",
  "coverage report",
]
cov = [
  "test-cov",
  "cov-report",
]

[[tool.hatch.envs.all.matrix]]
python = ["3.11"]

[tool.hatch.envs.types]
extra-dependencies = [
  "mypy>=1.0.0",
]
[tool.hatch.envs.types.scripts]
check = "mypy --install-types --non-interactive {args:src/validator tests}"

[tool.mypy]
plugins = "numpy.typing.mypy_plugin"

[tool.coverage.run]
source_pkgs = ["validator", "tests"]
branch = true
parallel = true
omit = [
  "src/validator/__about__.py",
]

[tool.coverage.paths]
validator = ["src/validator", "*/validator/src/validator"]
tests = ["tests", "*/validator/tests"]

[tool.coverage.report]
exclude_lines = [
  "no cov",
  "if __name__ == .__main__.:",
  "if TYPE_CHECKING:",
]

[tool.pytest.ini_options]
markers = [
	"focus",  # used in development to mark focused tests
]

[tool.ruff]
line-length = 120

